Move helper functions to relevant folders
