+++ functions/src/aqi-calculation/calculate-aqi.ts
@@ -1,3 +1,17 @@
+import {firestore, FieldValue, Timestamp} from '../admin';
+import {
+  bufferStatus,
+  defaultAqiBufferElement,
+  populateDefaultBuffer,
+  AqiBufferElement,
+  Pm25BufferElement,
+} from './buffer';
+import {CurrentReadingSensorData} from './util';
+import {
+  getCleanedAverages,
+  cleanedReadingsToNowCastPm25,
+} from './cleaned-reading';
+
 /**
  * Computes the AQI for PM 2.5 given the appropriate AQI breakpoints.
  * In most use cases, this function will be called from a function which knows those breakpoints.
@@ -37,7 +51,7 @@ function indexCalculation(
  * @returns the AQI for a given PM 2.5 concentration
  * @remarks The value is rounded to the nearest integer for valid AQI ranges. If the PM 2.5 concentration is less than zero, the reported AQI is negative infinity. If the PM 2.5 concentration is too high ("beyond the AQI" in EPA parlance), positive infinity is reported.
  */
-export function aqiFromPm25(pm25Concentration: number): number {
+function aqiFromPm25(pm25Concentration: number): number {
   // Source of bound values is Table 6 of the paper at
   // https://www.airnow.gov/sites/default/files/2018-05/aqi-technical-assistance-document-may2016.pdf
   /* eslint-disable no-magic-numbers */
@@ -114,3 +128,130 @@ export function aqiFromPm25(pm25Concentration: number): number {
   /* eslint-enable no-magic-numbers */
   return Math.round(aqi);
 }
+
+/**
+ * Using the Pm25Buffer, calculates the AQI if there is enough readings and
+ * updates the `current-readings` AQI data and AQI-validity status. Also updates
+ * the aqiBuffer in each sensor doc.
+ */
+async function calculateAqi(): Promise<void> {
+  const sensorList = (await firestore.collection('sensors').get()).docs;
+  const currentData = Object.create(null);
+  for (const sensorDoc of sensorList) {
+    const sensorDocData = sensorDoc.data();
+
+    // Data sent to the current-readings collection
+    // Initially the previous data's value or the default values
+    const currentSensorData: CurrentReadingSensorData = {
+      purpleAirId: sensorDocData.purpleAirId ?? '',
+      name: sensorDocData.name ?? '',
+      latitude: sensorDocData.latitude ?? NaN,
+      longitude: sensorDocData.latitude ?? NaN,
+      readingDocId: sensorDoc.id,
+      lastValidAqiTime: sensorDocData.lastValidAqiTime ?? null,
+      lastSensorReadingTime: sensorDocData.lastSensorReadingTime ?? null,
+      isActive: sensorDocData.isActive ?? true,
+      nowCastPm25: NaN,
+      aqi: NaN,
+      isValid: false,
+    };
+
+    if (currentSensorData.isActive) {
+      // Data used to calculate hourly averages
+      const pm25BufferStatus: bufferStatus =
+        sensorDocData.pm25BufferStatus ?? bufferStatus.DoesNotExist;
+      const pm25BufferIndex: number = sensorDocData.pm25BufferIndex ?? 0;
+      const pm25Buffer: Array<Pm25BufferElement> =
+        sensorDocData.pm25Buffer ?? [];
+
+      // Get cleaned hourly averages from the PM2.5 Buffer
+      const cleanedAverages = getCleanedAverages(
+        pm25BufferStatus,
+        pm25BufferIndex,
+        pm25Buffer
+      );
+
+      // NowCast formula from the EPA requires 2 out of the last 3 hours
+      // to be available
+      let validEntriesLastThreeHours = 0;
+      const THREE_HOURS = 3;
+      for (let i = 0; i < Math.min(THREE_HOURS, cleanedAverages.length); i++) {
+        if (!Number.isNaN(cleanedAverages[i])) {
+          validEntriesLastThreeHours++;
+        }
+      }
+
+      // If there's enough info, the sensor's data is updated
+      // If there isn't, we send the default AQI buffer element
+      const aqiBufferData: AqiBufferElement = defaultAqiBufferElement; // New data to add
+
+      // If there is not enough info, the sensor's status is not valid
+      const NOWCAST_RECENT_DATA_THRESHOLD = 2;
+      if (validEntriesLastThreeHours >= NOWCAST_RECENT_DATA_THRESHOLD) {
+        // If the calculated AQI is infinity, then the sensor value is not valid
+        // Only calculate the NowCast PM 2.5 value and the AQI if there is enough data
+        const nowCastPm25 = cleanedReadingsToNowCastPm25(cleanedAverages);
+        const aqi = aqiFromPm25(nowCastPm25);
+        if (
+          aqi !== Number.POSITIVE_INFINITY &&
+          aqi !== Number.NEGATIVE_INFINITY
+        ) {
+          currentSensorData.aqi = aqi;
+          currentSensorData.nowCastPm25 = nowCastPm25;
+          currentSensorData.isValid = true;
+          currentSensorData.lastValidAqiTime = Timestamp.fromDate(new Date());
+
+          aqiBufferData.aqi = currentSensorData.aqi;
+          aqiBufferData.timestamp = currentSensorData.lastValidAqiTime;
+        }
+      }
+
+      // Set data in map of sensor's PurpleAir ID to the sensor's most recent data
+      currentData[currentSensorData.purpleAirId] = currentSensorData;
+
+      // Update the AQI circular buffer for this element
+      const sensorDocRef = firestore.collection('sensors').doc(sensorDoc.id);
+      const status = sensorDocData.aqiBufferStatus ?? bufferStatus.DoesNotExist;
+
+      switch (status) {
+        case bufferStatus.Exists: {
+          // The buffer exists, proceed with normal update
+          const aqiBuffer: Array<AqiBufferElement> = sensorDocData.aqiBuffer;
+          aqiBuffer[sensorDocData.aqiBufferIndex] = aqiBufferData;
+
+          await sensorDocRef.update({
+            aqiBufferIndex:
+              (sensorDocData.aqiBufferIndex + 1) % aqiBuffer.length, // eslint-disable-line no-magic-numbers,
+            aqiBuffer: aqiBuffer,
+            lastValidAqiTime: currentSensorData.lastValidAqiTime,
+          });
+          break;
+        }
+        case bufferStatus.DoesNotExist: {
+          // Initialize populating the buffer with default values, don't update
+          // any values until the buffer status is Exists
+          await sensorDocRef.update({
+            aqiBufferStatus: bufferStatus.InProgress,
+            lastValidAqiTime: currentSensorData.lastValidAqiTime,
+          });
+          // This function updates the bufferStatus once the buffer has been
+          // fully initialized, which uses an additional write to the database
+          populateDefaultBuffer(true, sensorDoc.id);
+          break;
+        }
+        default:
+          // If the buffer status is In Progress we don't update the buffer
+          // because the buffer is still being initialized
+          break;
+      }
+    }
+  }
+
+  // Send AQI readings to current-reading to be displayed on the map
+  await firestore.collection('current-reading').doc('sensors').set({
+    lastUpdated: FieldValue.serverTimestamp(),
+    data: currentData,
+  });
+}
+
+export {calculateAqi};
